rename connect-form components to connection-form
