+++ packages/connection-form/src/index.ts
@@ -1,4 +1,4 @@
-import ConnectionForm from './components/connect-form';
+import ConnectionForm from './components/connection-form';
 import SaveConnectionModal from './components/save-connection-modal';
 import { useConnectionColor } from './hooks/use-connection-color';
 import { adjustConnectionOptionsBeforeConnect } from './hooks/use-connect-form';
